[#12] add mark parser method
